+++ code-review-github.el
@@ -85,7 +85,7 @@ https://github.com/wandersoncferreira/code-review#configuration"))
               nil
               :unpaginate t
               :headers code-review-github-diffheader
-              :auth 'code-review
+              :auth code-review-auth-login-marker
               :host code-review-github-host
               :callback callback
               :errorback #'code-review-github-errback)))
@@ -110,7 +110,7 @@ https://github.com/wandersoncferreira/code-review#configuration"))
               nil
               :unpaginate t
               :headers code-review-github-diffheader
-              :auth 'code-review
+              :auth code-review-auth-login-marker
               :host code-review-github-host
               :callback callback
               :errorback #'code-review-github-errback)))
@@ -497,7 +497,7 @@ https://github.com/wandersoncferreira/code-review#configuration"))
                   num)))
     (ghub-graphql query
                   nil
-                  :auth 'code-review
+                  :auth code-review-auth-login-marker
                   :host code-review-github-graphql-host
                   :callback callback
                   :errorback #'code-review-github-errback)))
@@ -521,7 +521,7 @@ Optionally ask for the FALLBACK? query."
                            (oref github owner)
                            (oref github repo))
                    nil
-                   :auth 'code-review
+                   :auth code-review-auth-login-marker
                    :host code-review-github-host
                    :noerror 'return)))
     (if (eq (car (-first-item resp)) 'message)
@@ -547,7 +547,7 @@ Optionally ask for the FALLBACK? query."
                                                    (a-get x 'name))
                                                  (oref github labels))
                                            []))
-             :auth 'code-review
+             :auth code-review-auth-login-marker
              :host code-review-github-host
              :errorback #'code-review-github-errback
              :callback (lambda (&rest _) (funcall callback)))))
@@ -559,7 +559,7 @@ Optionally ask for the FALLBACK? query."
                            (oref github owner)
                            (oref github repo))
                    nil
-                   :auth 'code-review
+                   :auth code-review-auth-login-marker
                    :host code-review-github-host
                    :noerror 'return)))
     (if (eq (car (-first-item resp)) 'message)
@@ -577,7 +577,7 @@ Optionally ask for the FALLBACK? query."
                      (oref github repo)
                      (oref github number))
              nil
-             :auth 'code-review
+             :auth code-review-auth-login-marker
              :host code-review-github-host
              :payload (a-alist 'assignees (-map (lambda (it)
                                                   (a-get it 'login))
@@ -592,7 +592,7 @@ Optionally ask for the FALLBACK? query."
                            (oref github owner)
                            (oref github repo))
                    nil
-                   :auth 'code-review
+                   :auth code-review-auth-login-marker
                    :host code-review-github-host
                    :noerror 'return)))
     (if (eq (car (-first-item resp)) 'message)
@@ -610,7 +610,7 @@ Optionally ask for the FALLBACK? query."
                       (oref github repo)
                       (oref github number))
               nil
-              :auth 'code-review
+              :auth code-review-auth-login-marker
               :host code-review-github-host
               :payload (a-alist 'milestone (a-get (oref github milestones) 'number))
               :errorback #'code-review-github-errback
@@ -627,7 +627,7 @@ Optionally ask for the FALLBACK? query."
                       (oref github repo)
                       (oref github number))
               nil
-              :auth 'code-review
+              :auth code-review-auth-login-marker
               :host code-review-github-host
               :payload (a-alist 'title (oref github title))
               :errorback #'code-review-github-errback
@@ -641,7 +641,7 @@ Optionally ask for the FALLBACK? query."
                       (oref github repo)
                       (oref github number))
               nil
-              :auth 'code-review
+              :auth code-review-auth-login-marker
               :host code-review-github-host
               :payload (a-alist 'body (a-get (oref github raw-infos) 'bodyText))
               :errorback #'code-review-github-errback
@@ -654,7 +654,7 @@ Optionally ask for the FALLBACK? query."
                     (oref github repo)
                     (oref github number))
             nil
-            :auth 'code-review
+            :auth code-review-auth-login-marker
             :host code-review-github-host
             :payload (a-alist 'commit_title (oref github title)
                               'commit_message (oref github description)
@@ -683,7 +683,7 @@ Optionally ask for the FALLBACK? query."
                        (oref github repo)
                        path)
                nil
-               :auth 'code-review
+               :auth code-review-auth-login-marker
                :host code-review-github-host
                :payload (a-alist 'content r))))
 
@@ -701,7 +701,7 @@ Optionally ask for the FALLBACK? query."
                          (oref github repo)
                          path)
                  nil
-                 :auth 'code-review
+                 :auth code-review-auth-login-marker
                  :host code-review-github-host)))
 
 (defclass code-review-submit-github-replies ()
@@ -727,7 +727,7 @@ Optionally ask for the FALLBACK? query."
                         nil
                         :payload (a-alist 'body (oref reply body))
                         :headers code-review-github-diffheader
-                        :auth 'code-review
+                        :auth code-review-auth-login-marker
                         :host code-review-github-host
                         :callback (lambda (&rest _))
                         :errorback #'code-review-github-errback)))
@@ -774,7 +774,7 @@ Optionally ask for the FALLBACK? query."
                        (oref pr repo)
                        (oref pr number))
                nil
-               :auth 'code-review
+               :auth code-review-auth-login-marker
                :payload payload
                :host code-review-github-host
                :errorback #'code-review-github-errback
@@ -807,7 +807,7 @@ Optionally ask for the FALLBACK? query."
                                            `((repo_owner . ,(oref github owner))
                                              (repo_name . ,(oref github repo))
                                              (cursor . ,cursor))
-                                           :auth 'code-review
+                                           :auth code-review-auth-login-marker
                                            :host code-review-github-graphql-host)))
             (let-alist graphql-res
               (setq has-next-page .data.repository.assignableUsers.pageInfo.hasNextPage
@@ -831,7 +831,7 @@ Optionally ask for the FALLBACK? query."
     (ghub-graphql query
                   `((input . ((pullRequestId . ,pr-id)
                               (userIds . ,user-ids))))
-                  :auth 'code-review
+                  :auth code-review-auth-login-marker
                   :host code-review-github-graphql-host
                   :callback (lambda (&rest _)
                               (message "Review requested successfully!")
@@ -844,7 +844,7 @@ Optionally ask for the FALLBACK? query."
                      (oref github owner)
                      (oref github repo))
              nil
-             :auth 'code-review
+             :auth code-review-auth-login-marker
              :host code-review-github-host
              :payload (a-alist 'body body 'title title)
              :errorback #'code-review-github-errback
@@ -906,7 +906,7 @@ Return the blob URL if BLOB? is provided."
                      (oref github repo)
                      (oref github number))
              nil
-             :auth 'code-review
+             :auth code-review-auth-login-marker
              :host code-review-github-host
              :payload (a-alist 'body comment-msg)
              :callback callback
@@ -919,7 +919,7 @@ Return the blob URL if BLOB? is provided."
                      (oref github repo)
                      (oref github number))
              nil
-             :auth 'code-review
+             :auth code-review-auth-login-marker
              :headers '(("Accept" . "application/vnd.github.v3+json"))
              :host code-review-github-host
              :payload (a-alist 'path (oref local-comment path)